cdk/cdk_requirements.txt
# CDK
aws-cdk-lib==2.147.3
aws-cdk.aws-appsync-alpha==2.10.0a0